+++ projects/igniteui-angular/migrations/update-18_2_0/index.spec.ts
@@ -1,7 +1,7 @@
 import * as path from 'path';
 
-import { EmptyTree } from '@angular-devkit/schematics';
 import { SchematicTestRunner, UnitTestTree } from '@angular-devkit/schematics/testing';
+import { setupTestTree } from '../common/setup.spec';
 
 const version = '18.2.0';
 
@@ -9,42 +9,13 @@ describe(`Update to ${version}`, () => {
     let appTree: UnitTestTree;
     const schematicRunner = new SchematicTestRunner('ig-migrate', path.join(__dirname, '../migration-collection.json'));
 
-    const configJson = {
-        projects: {
-            testProj: {
-                root: '/',
-                sourceRoot: '/testSrc',
-                architect: {
-                    build: {
-                        options: {
-                            styles: [
-                                "/testSrc/styles.scss"
-                            ] as (string | object)[]
-                        }
-                    }
-                }
-            }
-        },
-        schematics: {
-            '@schematics/angular:component': {
-                prefix: 'appPrefix'
-            }
-        }
-    };
-
     beforeEach(() => {
-        appTree = new UnitTestTree(new EmptyTree());
-        appTree.create('/angular.json', JSON.stringify(configJson));
-        appTree.create('/testSrc/styles.scss', `
-@use "mockStyles.scss";
-@forward something;
-`);
+        appTree = setupTestTree();
     });
 
     const migrationName = 'migration-40';
 
     it('Should replace deprecated `shouldGenerate` property with `autoGenerate`', async () => {
-        pending('set up tests for migrations through lang service');
         appTree.create(
             '/testSrc/appPrefix/component/grid-test.component.ts',
             `import { Component } from '@angular/core';
@@ -133,6 +104,22 @@ describe(`Update to ${version}`, () => {
             `$custom-badge: badge-theme(
                 $text-color: red
             );`
+        );
+    });
+
+    it('should replace QueryBuilder deprecated property `fields` with `entities`', async () => {
+        appTree.create(`/testSrc/appPrefix/component/test.component.html`,
+        `
+        <igx-query-builder [fields]="[{ field: 'ID', dataType: 'number' }, { field: 'Name', dataType: 'string' }]"></igx-query-builder>
+        `
+        );
+
+        const tree = await schematicRunner.runSchematic(migrationName, { shouldInvokeLS: false }, appTree);
+
+        expect(tree.readContent('/testSrc/appPrefix/component/test.component.html')).toEqual(
+        `
+        <igx-query-builder [entities]="[{ name: '', fields: [{ field: 'ID', dataType: 'number' }, { field: 'Name', dataType: 'string' }]}]"></igx-query-builder>
+        `
         );
     });
 });