+++ projects/igniteui-angular/migrations/update-18_2_0/index.ts
@@ -3,12 +3,18 @@ import type {
     SchematicContext,
     Tree
 } from '@angular-devkit/schematics';
-import { UpdateChanges } from '../common/UpdateChanges';
+import { BoundPropertyObject, InputPropertyType, UpdateChanges } from '../common/UpdateChanges';
 
 const version = '18.2.0';
 
 export default (): Rule => async (host: Tree, context: SchematicContext) => {
     context.logger.info(`Applying migration for Ignite UI for Angular to version ${version}`);
     const update = new UpdateChanges(__dirname, host, context);
+
+    update.addValueTransform('fields_to_entities', (args: BoundPropertyObject): void => {
+        args.bindingType = InputPropertyType.EVAL;
+        args.value = `[{ name: '', fields: ${args.value}}]`;
+    });
+
     update.applyChanges();
 };