+++ projects/igniteui-angular/src/lib/data-operations/filtering-expression.interface.ts
@@ -1,5 +1,4 @@
 import { IFilteringOperation } from './filtering-condition';
-import { Serializable } from 'node:child_process';
 import { IExpressionTree } from './filtering-expressions-tree';
 
 /* mustCoerceToInt */
@@ -16,8 +15,8 @@ export enum FilteringLogic {
 export declare interface IFilteringExpression {
     fieldName: string;
     condition?: IFilteringOperation;
-    conditionName: string;
-    searchVal?: Serializable;
+    conditionName?: string;
+    searchVal?: any;
     searchTree?: IExpressionTree;
     ignoreCase?: boolean;
 }