+++ CHANGELOG.md
@@ -14,7 +14,7 @@ All notable changes for each version of this project will be documented in this 
     - Added support for showing/hiding the indicator controls (dots). Can be configured via the `indicators` property. Defaults to `true`.
 
 - `IFilteringExpression`
-    - **Breaking Change** There is a new `conditionName` property which is required. This would generally be equal to the old `condition.name`.
+    - A new optional property called `conditionName` has been introduced. This would generally be equal to the existing `condition.name`.
     - A new optional property called `isNestedQuery` has been introduced. It's used to indicate whether the condition leads to a sub-query creation.
 
 #### Scrollbar: New CSS variables